refactor(proxy): break circular deps & unify CONTEXT_WINDOW_OPTIONS
Address remaining pre-existing issues from code review:

- Extract transport/shared.ts to break proxy-core<->http<->stream cycles
  (3 circular deps eliminated per fallow). Move buildUpstreamUrl,
  findPathOverlap, KNOWN_API_SUFFIXES, UpstreamRequestOptions,
  _transportInternals, buildRequestOptions, BuildHeadersFn,
  TransportCallOpts to shared. http.ts re-exports _transportInternals
  for test mock compatibility; stream.ts & iteration-setup.ts import
  from shared. proxy-core.ts no longer referenced by transport layer.
- Unify CONTEXT_WINDOW_OPTIONS: single source in quick-setup/types.ts
  (number value); useProviderForm re-exports instead of redefining with
  string value (duplicate export + type divergence eliminated).
  ModelCapabilitiesEditor SelectItem value wrapped with String().
- ModelCard.vue: add missing Button import (Trash icon button was
  unresolvable at runtime).
- Update test mock paths: stream-buffering-opt now mocks shared.js
  (follows dependency migration); build-upstream-url imports from shared.

Verified: tsc 0 error, ESLint 0 warn, 1834 tests pass, fallow reports
0 circular deps & 0 duplicate exports.
