+++ router/src/proxy/handler/iteration-setup.ts
@@ -11,7 +11,8 @@ import type { ProxyErrorFormatter } from "../proxy-core.js";
 import type { FormatAdapter } from "../format/types.js";
 import type { FormatRegistry } from "../format/registry.js";
 import { sanitizeHeadersForLog } from "../proxy-logging.js";
-import { buildUpstreamHeaders, buildUpstreamUrl } from "../proxy-core.js";
+import { buildUpstreamHeaders } from "../proxy-core.js";
+import { buildUpstreamUrl } from "../transport/shared.js";
 import { getModelTimeouts } from "../../db/providers.js";
 import { buildTransportFn } from "../transport/transport-fn.js";
 import { parseModels } from "../../config/model-context.js";
